refactor(NestedList): package 이동 (domain.collections -> utils)
